+++ CHANGELOG.rst
@@ -2,10 +2,15 @@
 Changelog
 =========
 
+Version 2.9.0 (2023-03-18)
+==========================
+
+Compatible with bthome-ble 2.9.0. Adds water sensor.
+
 Version 2.8.0 (2023-03-18)
 ==========================
 
-* Compatible with bthome-ble 2.8.0
+Compatible with bthome-ble 2.8.0. Adds gas, energy, volume sensor.
 
 Version 2.7.0 (2023-02-26)
 ==========================